backend auth: small cleanup
