+++ src/backend/src/auth/auth.service.ts
@@ -128,13 +128,7 @@ export class AuthService {
     if (!data)
       throw new ForbiddenException('Could not fetch user with 42 OAuth2 API');
     const user = new OAuthUserDto();
-    for (const key of [
-      'login',
-      'email',
-      'image_url',
-      // 'first_name',
-      // 'last_name',
-    ]) {
+    for (const key of ['login', 'email', 'image_url']) {
       user[key] = data[key];
     }
     const errors = await validate(user, {
