+++ src/backend/src/auth/decorator/index.ts
@@ -1,2 +1,1 @@
-export * from './match-property.decorator';
 export * from './user.decorator';
