--- src/backend/src/auth/decorator/match-property.decorator.ts
@@ -1,34 +1,0 @@
-import {
-  registerDecorator,
-  ValidationArguments,
-  ValidationOptions,
-  ValidatorConstraint,
-  ValidatorConstraintInterface,
-} from 'class-validator';
-
-export function MatchProperty(
-  property: string,
-  validationOptions?: ValidationOptions,
-) {
-  return (object: any, propertyName: string) => {
-    registerDecorator({
-      target: object.constructor,
-      propertyName,
-      options: validationOptions,
-      constraints: [property],
-      validator: MatchConstraint,
-    });
-  };
-}
-
-@ValidatorConstraint()
-export class MatchConstraint implements ValidatorConstraintInterface {
-  validate(value: any, args: ValidationArguments) {
-    const [relatedPropertyName] = args.constraints;
-    const relatedValue = (args.object as any)[relatedPropertyName];
-    return value === relatedValue;
-  }
-}
-
-// Credit to: https://stackoverflow.com/questions/60451337/password-confirmation-in-typescript-with-class-validator
-// Doc on custom validation decorators: https://github.com/typestack/class-validator#custom-validation-decorators
