+++ src/backend/src/auth/dto/index.ts
@@ -1,1 +1,2 @@
-export * from './auth.dto';
+export * from './local-auth.dto';
+export * from './oauth.dto';
