+++ src/backend/src/auth/dto/local-auth.dto.ts
@@ -1,0 +1,33 @@
+import {
+  IsEmail,
+  IsNotEmpty,
+  IsString,
+  MaxLength,
+  MinLength,
+} from 'class-validator';
+
+export class LocalSignupDto {
+  @IsString()
+  @IsNotEmpty()
+  @MaxLength(42)
+  username: string;
+
+  @IsEmail()
+  @IsNotEmpty()
+  email: string;
+
+  @IsString()
+  @MinLength(4)
+  @MaxLength(42)
+  password: string;
+}
+
+export class LocalSigninDto {
+  @IsString()
+  @IsNotEmpty()
+  username: string;
+
+  @IsString()
+  @IsNotEmpty()
+  password: string;
+}
