+++ src/backend/src/auth/dto/oauth.dto.ts
@@ -1,0 +1,36 @@
+import {
+  IsEmail,
+  IsNotEmpty,
+  IsString,
+  IsUrl,
+  MaxLength,
+} from 'class-validator';
+
+/*
+ * @brief Partial user infos fetched with 42's API
+ */
+export class OAuthUserDto {
+  @IsString()
+  @IsNotEmpty()
+  login: string;
+
+  @IsEmail()
+  @IsNotEmpty()
+  email: string;
+
+  @IsUrl()
+  image_url: string;
+}
+
+/*
+ * @brief Data sent in signup in OAuth mode
+ */
+export class OAuthSignUpDto {
+  @IsString()
+  @IsNotEmpty()
+  @MaxLength(42)
+  username: string;
+
+  @IsString()
+  jwt: string;
+}
